# openapi/target/classes/openapi.yml
openapi: 3.0.3
info:
  title: API oPROsita
  description: API REST para gestión de grupos, usuarios, contenido y simulacros
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /contenido/{id}:
    delete:
      tags:
        - contenido
      operationId: deleteContenido
      summary: Eliminar un contenido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Contenido eliminado correctamente
    put:
      tags:
        - contenido
      operationId: updateContenido
      summary: Actualizar un contenido
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContenidoItem'
      responses:
        '200':
          description: Contenido actualizado
...
  /chat/mensajes:
    get:
      tags:
        - chat
      operationId: getMensajes
      summary: Obtener mensajes
      parameters:
        - in: query
          name: remitente
          required: false
          schema:
            type: integer
        - in: query
          name: destinatario
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Lista de mensajes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mensaje'
    post:
      tags:
        - chat
      operationId: enviarMensaje
      summary: Enviar un mensaje
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mensaje'
      responses:
        '201':
          description: Mensaje enviado
  /chat/conversaciones:
    get:
      tags:
        - chat
      operationId: getConversaciones
      summary: Obtener lista de conversaciones
      parameters:
        - in: query
          name: usuarioId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de conversaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversacion'
components:
  schemas:
...
    ContenidoItem:
      type: object
      properties:
        id:
          type: integer
        texto:
          type: string
        tipoContenido:
          type: string
          enum: [temas, programacion, practico]
        alumnoId:
          type: integer
          nullable: true
        grupoId:
          type: integer
        mes:
          type: string
        archivoId:
          type: integer
          nullable: true
...
    Mensaje:
      type: object
      properties:
        id:
          type: integer
        remitente:
          type: integer
          description: ID del usuario remitente
        destinatario:
          type: integer
          description: ID del usuario destinatario
        contenido:
          type: string
        leido:
          type: boolean
        fechaEnvio:
          type: string
          format: date-time
    Conversacion:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        otroUsuarioId:
          type: integer
        ultimoMensaje:
          type: string
        fechaUltimoMensaje:
          type: string
          format: date-time
        noLeidos:
          type: integer
...